added model for skill that will be used to seed skills. renamed interfaces/models to definedmodels for easier navigation to not be confused with models.
