+++ src/interfaces/definedmodels/ClassInterface.ts
