+++ src/interfaces/definedmodels/InventoryInterface.ts
