+++ src/interfaces/definedmodels/ItemInterface.ts
