+++ src/interfaces/definedmodels/PartyInterfaces.ts
