+++ src/interfaces/definedmodels/PlayerInterfaces.ts
