+++ src/interfaces/definedmodels/SkillsInterface.ts
